chore: refactor folder and task context separation
Moved selectedList and selectedFolderId state management from TasksProvider to FoldersProvider for clearer separation of concerns. Updated components to use useFolders for list and folder selection, and adjusted context usage throughout the app. Initialization of folders and lists now occurs in Home.jsx, and providers are nested in main.jsx to ensure proper context availability.
