+++ src/components/Sidebar/Logo.jsx
@@ -1,12 +1,20 @@
-import { useTasks } from "../../hooks/useTasks";
+import { useFolders } from "../../hooks/useFolders";
+import { useNavigate } from "react-router-dom";
 
 export const Logo = () => {
-  const { unSelectList } = useTasks();
+  const { unSelectList } = useFolders();
+
+  const navigate = useNavigate();
+
+  const clearLists = () => {
+    unSelectList();
+    navigate("/");
+  };
 
   return (
     <div className="flex gap-1 items-center mb-5">
       <div
-        onClick={unSelectList}
+        onClick={clearLists}
         className="flex items-center gap-1 cursor-pointer"
       >
         <img width="20px" height="20px" src="/taskify.png" alt="Taskify Logo" />
